Various Bug fixes
When the battle starts, nothing exist on screen, and you need to click once to spawn everything, and click again to spawn the ball. The shoot ball script has been removed and the spawning of the player's ball in now handled in paddle. A small icon has been added to allow the player to see where the ball/laser/other objects spawn from on the paddle based on tester feedback. The combo meter is no longer maxed at 10, but the speed boost from it is.