All the animations are now in
I put all the animations in. Now I just need to figure out their sizes, espically an oddity in the Shockwave Attack where one sprite is much smaller than the others (It's not part of the spreadsheet and is instead its own pic so that's most likely the reason). I also still need to figure out how the transitions between animations work so that I can properly implement the animations to their corresponding attacks.